feat: refactor Docker and Podman tests to use common test utilities
